requirements.txt
psycopg2-binary
django-cors-headers
asyncpg